examples/02-dynamic-mcp/agent/pyproject.toml
[project]
name = "mcp-deployer"
version = "0.1.0"
description = "BYO ADK agent that discovers, deploys, and uses MCP servers on-demand"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.25.0",
    "mcp>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
